Refactored test cases to be in separate files
